Final Project
Modify the view-data page to be an AngularJS page. - 20 Points
Add the ability to sort the table by a "type" - 35 points
The type should be something that is used by many records
For example:  In the library repository you could use author or publisher as the type.  Avoid using the name (of a game, book, restaurant, etc.) field
Add a "get by type" listener to your server
Add the ability to update a record- 35 points
This can be done on the repository table page (as we did in the Harry Potter example) or on a new page.
Use the PUT method
Add a PUT listener to your server
Keep it pretty (CSS) - 10 Points

Not super happy with the styles on the write-data page, but I'm so tired of dealing with it and I know that the page doesn't look amazing as is.